move pydantic tests from test_annotated_fields
