+++ tests/test_pydantic_factory.py
@@ -1,5 +1,6 @@
 import sys
 from collections import Counter, deque
+from dataclasses import dataclass
 from datetime import date, datetime, time, timedelta
 from decimal import Decimal
 from ipaddress import (
@@ -11,11 +12,11 @@ from ipaddress import (
     IPv6Network,
 )
 from pathlib import Path
-from typing import Callable, Dict, List, Optional, Set, Tuple, Union
+from typing import Callable, Dict, List, Literal, Optional, Set, Tuple, Union
 from uuid import UUID
 
 import pytest
-from annotated_types import Ge, MinLen
+from annotated_types import Ge, Le, LowerCase, MinLen, UpperCase
 from typing_extensions import Annotated, TypeAlias
 
 import pydantic
@@ -60,6 +61,7 @@ from pydantic import (
     ValidationError,
 )
 
+from polyfactory.factories import DataclassFactory
 from polyfactory.factories.pydantic_factory import _IS_PYDANTIC_V1, ModelFactory
 from tests.models import PetFactory
 
@@ -439,3 +441,83 @@ def test_optional_type(allow_none: bool) -> None:
         __allow_none_optionals__ = allow_none
 
     assert AFactory.build()
+
+
+def test_discriminated_unions() -> None:
+    class BasePet(BaseModel):
+        name: str
+
+    class BlackCat(BasePet):
+        pet_type: Literal["cat"]
+        color: Literal["black"]
+
+    class WhiteCat(BasePet):
+        pet_type: Literal["cat"]
+        color: Literal["white"]
+
+    class Dog(BasePet):
+        pet_type: Literal["dog"]
+
+    class Owner(BaseModel):
+        pet: Annotated[
+            Union[Annotated[Union[BlackCat, WhiteCat], Field(discriminator="color")], Dog],
+            Field(discriminator="pet_type"),
+        ]
+        name: str
+
+    class OwnerFactory(ModelFactory):
+        __model__ = Owner
+
+    assert OwnerFactory.build()
+
+
+def test_predicated_fields() -> None:
+    @dataclass
+    class PredicatedMusician:
+        name: Annotated[str, UpperCase]
+        band: Annotated[str, LowerCase]
+
+    class PredicatedMusicianFactory(DataclassFactory):
+        __model__ = PredicatedMusician
+
+    assert PredicatedMusicianFactory.build()
+
+
+def test_tuple_with_annotated_constraints() -> None:
+    class Location(BaseModel):
+        long_lat: Tuple[Annotated[float, Ge(-180), Le(180)], Annotated[float, Ge(-90), Le(90)]]
+
+    class LocationFactory(ModelFactory[Location]):
+        __model__ = Location
+
+    assert LocationFactory.build()
+
+
+def test_optional_tuple_with_annotated_constraints() -> None:
+    class Location(BaseModel):
+        long_lat: Union[Tuple[Annotated[float, Ge(-180), Le(180)], Annotated[float, Ge(-90), Le(90)]], None]
+
+    class LocationFactory(ModelFactory[Location]):
+        __model__ = Location
+
+    assert LocationFactory.build()
+
+
+def test_legacy_tuple_with_annotated_constraints() -> None:
+    class Location(BaseModel):
+        long_lat: Tuple[Annotated[float, Ge(-180), Le(180)], Annotated[float, Ge(-90), Le(90)]]
+
+    class LocationFactory(ModelFactory[Location]):
+        __model__ = Location
+
+    assert LocationFactory.build()
+
+
+def test_legacy_optional_tuple_with_annotated_constraints() -> None:
+    class Location(BaseModel):
+        long_lat: Union[Tuple[Annotated[float, Ge(-180), Le(180)], Annotated[float, Ge(-90), Le(90)]], None]
+
+    class LocationFactory(ModelFactory[Location]):
+        __model__ = Location
+
+    assert LocationFactory.build()
